Messaging and refactor of pipeline simulators. (#295)
* Refactor filter code.

* Moved pipeline_messaging

* Typing updates.

* Updated documentation.

* Formatting.
