+++ epymorph/forecasting/pipeline.py
@@ -3,29 +3,23 @@ Simulators for performing multi-realization simulations such as forecasting and
 state/parameter fitting.
 """
 
-import dataclasses
 import datetime
 from abc import ABC, abstractmethod
 from dataclasses import dataclass, replace
-from typing import Generic, Mapping, Self, Sequence, TypeVar
+from typing import Mapping, Self
 
 import numpy as np
 import pandas as pd
 from numpy.typing import NDArray
-from typing_extensions import override
 
-from epymorph.adrio.adrio import ADRIO, ValueT
 from epymorph.attribute import NamePattern
 from epymorph.compartment_model import (
     CompartmentDef,
-    QuantityAggregation,
-    QuantitySelection,
     QuantityStrategy,
     TransitionDef,
 )
 from epymorph.data_type import SimDType
 from epymorph.forecasting.dynamic_params import ParamFunctionDynamics, Prior
-from epymorph.forecasting.likelihood import Likelihood
 from epymorph.forecasting.munge_realizations import (
     ParameterSelector,
     ParameterStrategy,
@@ -34,21 +28,15 @@ from epymorph.forecasting.munge_realizations import (
     RealizationSelector,
 )
 from epymorph.geography.scope import GeoAggregation, GeoSelection
-from epymorph.initializer import Explicit
 from epymorph.rume import RUME
-from epymorph.simulation import Context
-from epymorph.simulator.basic.basic_simulator import BasicSimulator
 from epymorph.simulator.basic.output import Output
 from epymorph.time import (
     Dim,
-    GroupKeyType,
     TimeAggregation,
     TimeFrame,
-    TimeGrouping,
     TimeSelection,
 )
-from epymorph.tools.data import mask, munge
-from epymorph.util import DateValueType
+from epymorph.tools.data import mask
 
 
 @dataclass(frozen=True)
@@ -399,1070 +387,6 @@ class PipelineSimulator(ABC):
         """
         Run the simulation.
         """
-
-
-def _initialize_compartments_and_params(
-    rume: RUME,
-    unknown_params: Mapping[NamePattern, UnknownParam],
-    num_realizations: int,
-    initial_values: NDArray[SimDType] | None,
-    rng: np.random.Generator,
-) -> tuple[NDArray, Mapping[NamePattern, NDArray]]:
-    """
-    Parameters
-    ----------
-    rume :
-        The RUME.
-    unknown_params :
-        The unknown parameters. The initial parameter values are based on the prior of
-        each UnknownParam.
-    num_realizations :
-        The number of realizations.
-    initial_values :
-        An array of shape (R, N, C) containing the initial compartment values.
-        If None then the RUME's intitializer is used to generate a single initial value
-        for each compartment.
-    rng :
-        The random number generator.
-
-    Returns
-    -------
-    :
-        The initial compartment values and the initial paramter values.
-    """
-    num_compartments = rume.ipm.num_compartments
-    current_params = {
-        k: np.zeros(shape=(num_realizations, rume.scope.nodes), dtype=np.float64)
-        for k in unknown_params.keys()
-    }
-    for i_realization in range(num_realizations):
-        for name in current_params.keys():
-            prior = unknown_params[name].prior
-            if isinstance(prior, Prior):
-                current_params[name][i_realization, ...] = prior.sample(
-                    size=(rume.scope.nodes,), rng=rng
-                )
-            else:
-                current_params[name][i_realization, ...] = prior[i_realization, ...]
-
-    current_compartments = np.zeros(
-        shape=(num_realizations, rume.scope.nodes, num_compartments), dtype=np.int64
-    )
-    if initial_values is not None:
-        current_compartments = initial_values.copy()
-    else:
-        params_temp = {**rume.params}
-        for i_realization in range(num_realizations):
-            for name in current_params.keys():
-                params_temp[name] = current_params[name][i_realization, ...]
-            rume_temp = dataclasses.replace(rume, params=params_temp)
-            data = rume_temp.evaluate_params(rng=rng)
-            current_compartments[i_realization, ...] = rume.initialize(
-                data=data, rng=rng
-            )
-    return current_compartments, current_params
-
-
-@dataclass(frozen=True)
-class _SimulateRealizationsResult:
-    compartments: NDArray[SimDType]
-    """
-    An array of shape (R, S, N, C) where R is the number of realizations, S is the
-    number of tau steps, N is the number of nodes, and C is the number of compartments.
-    Contains the compartment values of each realization over time.
-    """
-
-    events: NDArray[SimDType]
-    """
-    An array of shape (R, S, N, E) where R is the number of realizations, S is the
-    number of tau steps, N is the number of nodes, and E is the number of events.
-    Contains the event values of each realization over time.
-    """
-
-    estimated_params: Mapping[NamePattern, NDArray]
-    """
-    A dictionary containing arrays of shape (R, T, N) where R is the number of
-    realizations, T is the number of days, and N is the number of nodes. Each
-    array contains the values of the unknown parameters over time.
-    """
-
-    predictions: list[NDArray]
-    """
-    A list of arrays containing the predicted observation corresponding to each
-    realization.
-    """
-
-
-def _simulate_realizations(
-    rume_template: RUME,
-    override_time_frame: TimeFrame,
-    num_realizations: int,
-    initial_values: NDArray[SimDType],
-    unknown_params: Mapping[NamePattern, UnknownParam],
-    param_values: Mapping[NamePattern, NDArray[np.float64]],
-    geo: GeoSelection | GeoAggregation,
-    time: TimeSelection | TimeAggregation,
-    quantity: QuantitySelection | QuantityAggregation,
-    rng: np.random.Generator,
-) -> _SimulateRealizationsResult:
-    """
-    A helper function used by ForecastSimulator and ParticleFilterSimulator to propagate
-    realizations forward in time.
-
-    Parameters
-    ----------
-    rume_template :
-        A partial RUME which will be filled out with the compartment and parameter
-        values of each realization.
-    override_time_frame :
-        An override of the time_frame of the RUME template.
-    num_realizations :
-        The number of realizations.
-    initial_values :
-        An array of shape (R, N, C) containing the compartment values of each
-        realization.
-    unknown_params :
-        A dictionary containing the unknown parameters which are allowed to vary across
-        realizations. The prior field of each UnknownParam is ingnored in favor of
-        param_values.
-    param_values :
-        A dictionary containing the current parameter values.
-    geo :
-        The geo strategy for munging the output.
-    time :
-        The time strategy for munging the output.
-    quantity :
-        The quantity strategy for munging the output.
-    rng :
-        The random number generator.
-
-    Returns
-    -------
-    :
-        The result of the simulation.
-    """
-    num_steps = override_time_frame.days * rume_template.num_tau_steps
-
-    compartments = np.zeros(
-        shape=(
-            num_realizations,
-            num_steps,
-            rume_template.scope.nodes,
-            rume_template.ipm.num_compartments,
-        ),
-        dtype=np.int64,
-    )
-    events = np.zeros(
-        shape=(
-            num_realizations,
-            num_steps,
-            rume_template.scope.nodes,
-            rume_template.ipm.num_events,
-        ),
-        dtype=np.int64,
-    )
-    estimated_params = {
-        k: np.zeros(
-            shape=(
-                num_realizations,
-                override_time_frame.days,
-                rume_template.scope.nodes,
-            ),
-            dtype=np.float64,
-        )
-        for k in unknown_params.keys()
-    }
-
-    current_predictions = []
-    for i_realization in range(num_realizations):
-        params_temp = {**rume_template.params}
-        for name in unknown_params.keys():
-            params_temp[name] = unknown_params[name].dynamics.with_initial(
-                param_values[name][i_realization, ...]
-            )
-        rume_temp = dataclasses.replace(
-            rume_template, params=params_temp, time_frame=override_time_frame
-        )
-
-        # Evaluate the params.
-        data = rume_temp.evaluate_params(rng=rng)
-        rume_temp = dataclasses.replace(
-            rume_temp,
-            params={k.to_pattern(): v for k, v in data.to_dict().items()},
-        )
-
-        # Add the compartment values.
-        strata_temp = [
-            dataclasses.replace(
-                stratum,
-                init=Explicit(
-                    initials=initial_values[i_realization, ...][
-                        ..., rume_template.compartment_mask[stratum.name]
-                    ]
-                ),
-            )
-            for stratum in rume_template.strata
-        ]
-        rume_temp = dataclasses.replace(rume_temp, strata=strata_temp)
-
-        sim = BasicSimulator(rume_temp)
-        out_temp = sim.run(rng_factory=lambda: rng)
-
-        temp_time_agg = out_temp.rume.time_frame.select.all().agg()
-        new_time_agg = dataclasses.replace(
-            temp_time_agg,
-            aggregation=time.aggregation,
-        )
-
-        predicted_df = munge(
-            out_temp,
-            geo=geo,
-            time=new_time_agg,
-            quantity=quantity,
-        )
-
-        prediction = predicted_df.drop(["geo", "time"], axis=1)
-
-        current_predictions.append(prediction.to_numpy())
-
-        compartments[i_realization, ...] = out_temp.compartments
-        events[i_realization, ...] = out_temp.events
-
-        for name in estimated_params.keys():
-            matching_names = [x for x in data.raw_values if name.match(x)]
-            # We know that all the matching names have identical values.
-            estimated_params[name][i_realization, ...] = data.raw_values[
-                matching_names[0]
-            ]
-
-    return _SimulateRealizationsResult(
-        compartments=compartments,
-        events=events,
-        estimated_params=estimated_params,
-        predictions=current_predictions,
-    )
-
-
-@dataclass(frozen=True)
-class ForecastSimulator(PipelineSimulator):
-    """
-    PipelineSimulator for running a multi-realization forecast of a RUME and any
-    unknown parameters.
-
-    Parameters
-    ----------
-    config :
-        Contains the rume, number of realizations, initial compartment values, and
-        dictionary of unknown parameters.
-    """
-
-    config: PipelineConfig
-    """The configuration for the simulator."""
-
-    @override
-    def run(self, rng: np.random.Generator) -> PipelineOutput:
-        """
-        Runs the multi-realization forecast.
-
-        Parameters
-        ----------
-        rng : np.random.Generator
-            The random number generator.
-
-        Returns
-        -------
-        :
-            The output of the forecast.
-        """
-
-        # Precompute ADRIO values.
-        context = Context.of(
-            scope=self.rume.scope, time_frame=self.rume.time_frame, rng=rng
-        )
-        new_params = {}
-        for name, param in self.rume.params.items():
-            if isinstance(param, ADRIO):
-                new_params[name] = param.with_context_internal(context).evaluate()
-            else:
-                new_params[name] = param
-        rume = dataclasses.replace(
-            self.rume,
-            params=new_params,
-        )
-
-        # Initialize the initial compartment and parameter values.
-        current_compartments, current_params = _initialize_compartments_and_params(
-            rume, self.unknown_params, self.num_realizations, self.initial_values, rng
-        )
-        initial = current_compartments
-
-        result = _simulate_realizations(
-            rume_template=rume,
-            override_time_frame=rume.time_frame,
-            num_realizations=self.num_realizations,
-            initial_values=current_compartments,
-            unknown_params=self.unknown_params,
-            param_values=current_params,
-            geo=rume.scope.select.all(),
-            time=rume.time_frame.select.all(),
-            quantity=rume.ipm.select.all(),
-            rng=rng,
-        )
-
-        return PipelineOutput(
-            simulator=self,
-            final_compartments=result.compartments[:, -1, ...],
-            final_params={
-                k: result.estimated_params[k][:, -1, ...]
-                for k in result.estimated_params.keys()
-            },
-            compartments=result.compartments,
-            events=result.events,
-            initial=initial,
-            estimated_params=result.estimated_params,
-        )
-
-
-###################
-# Abstract Filter #
-###################
-
-
-@dataclass(frozen=True)
-class ModelLink:
-    """
-    Contains the information needed to subselect, group, and aggregate model output in
-    order to compare it to observed data.
-
-    Parameters
-    ----------
-    geo :
-        The strategy for processing the geo-axis of a single realization.
-    time :
-        The strategy for processing the time-axis of a single realization.
-    quantity :
-        The strategy for processing the quantity-axis of a single realization.
-    """
-
-    geo: GeoSelection | GeoAggregation
-    """The strategy for processing the geo-axis of a single realization."""
-
-    time: TimeSelection | TimeAggregation
-    """The strategy for processing the time-axis of a single realization."""
-
-    quantity: QuantitySelection | QuantityAggregation
-    """The strategy for processing the quantity-axis of a single realization."""
-
-
-@dataclass(frozen=True)
-class Observations(Generic[ValueT]):
-    """
-    The observed data used to estimate the state and parameters of a simulation.
-
-    Parameters
-    ----------
-    source :
-        The source of the observations which results in an array of shape (A, N) or
-        (A, N, Q) array where A is a time axis of unknown length, N is the number of
-        nodes, and Q are quantities. The entries of the array are date-value pairs.
-    model_link :
-        Contains the information needed to generate a predicted observation from a
-        single realization of a simulation.
-    likelihood :
-        Contains a likelihood function used to compare a predicted observation to
-        observed data.
-    """
-
-    source: ADRIO[DateValueType, ValueT] | NDArray[DateValueType]
-    """
-    The source of the observations which results in an  array of observed values.
-    """
-
-    model_link: ModelLink
-    """
-    Contains the information needed to generate a predicted observation from a
-    single realization of a simulation.
-    """
-
-    likelihood: Likelihood
-    """
-    Contains a likelihood function used to compare a predicted observation to observed
-    data.
-    """
-
-    def _get_observations_array(self, context: Context) -> NDArray[DateValueType]:
-        """
-        Get the arbitrary by node structured array of date value pairs.
-
-        Parameters
-        ----------
-        context :
-            The context for evaluating an ADRIO source of the observations.
-        """
-        if isinstance(self.source, ADRIO):
-            inspect = self.source.with_context_internal(context).inspect()
-            return inspect.result
-        else:
-            return self.source
-
-
-_FilterContextT = TypeVar("_FilterContextT")
-
-
-@dataclass(frozen=True)
-class _FilterUpdateResult(Generic[_FilterContextT]):
-    """"""
-
-    current_compartments: NDArray
-    current_params: Mapping[NamePattern, NDArray]
-    posterior_values: np.ndarray
-    filter_context: _FilterContextT
-
-
-@dataclass(frozen=True)
-class _RunFilterResult(Generic[_FilterContextT]):
-    final_compartments: NDArray[SimDType]
-    final_params: Mapping[NamePattern, NDArray[np.float64]]
-    compartments: NDArray[SimDType]
-    events: NDArray[SimDType]
-    initial: NDArray[SimDType]
-    estimated_params: Mapping[NamePattern, NDArray[np.float64]]
-    posterior_values: NDArray
-    filter_context: _FilterContextT
-
-
-@dataclass(frozen=True)
-class FilterOutput(PipelineOutput):
-    posterior_values: NDArray[np.float64]
-
-
-@dataclass(frozen=True)
-class FilterSimulator(PipelineSimulator, Generic[_FilterContextT]):
-    config: PipelineConfig
-    """The configuration for the simulation."""
-
-    observations: Observations
-    """The observations."""
-
-    @abstractmethod
-    def _initialize_filter_context(
-        self,
-    ) -> _FilterContextT:
-        """"""
-
-    @abstractmethod
-    def run(self, rng: np.random.Generator) -> FilterOutput:
-        """"""
-
-    @abstractmethod
-    def _update(
-        self,
-        current_observation: NDArray,
-        current_predictions: NDArray,
-        prior_compartments: NDArray,
-        prior_params: Mapping[NamePattern, NDArray],
-        filter_context: _FilterContextT,
-        rng: np.random.Generator,
-    ) -> _FilterUpdateResult:
-        """"""
-
-    def _run_filter(self, rng: np.random.Generator) -> _RunFilterResult:
-        """
-        Run the particle filter simulation.
-
-        Parameters
-        ----------
-        rng : np.random.Generator
-            The random number generator.
-
-        Returns
-        -------
-        :
-            The output of the particle filter.
-        """
-        # Precompute ADRIO values.
-        context = Context.of(
-            scope=self.rume.scope, time_frame=self.rume.time_frame, rng=rng
-        )
-        new_params = {}
-        for name, param in self.rume.params.items():
-            if isinstance(param, ADRIO):
-                new_params[name] = param.with_context_internal(context).evaluate()
-            else:
-                new_params[name] = param
-
-        rume = dataclasses.replace(
-            self.rume,
-            params=new_params,
-        )
-
-        # Dimension of the system.
-        num_steps = rume.time_frame.days * rume.num_tau_steps
-
-        # Allocate return values.
-        initial = np.zeros(
-            shape=(self.num_realizations, rume.scope.nodes, rume.ipm.num_compartments),
-            dtype=np.int64,
-        )
-        compartments = np.zeros(
-            shape=(
-                self.num_realizations,
-                num_steps,
-                rume.scope.nodes,
-                rume.ipm.num_compartments,
-            ),
-            dtype=np.int64,
-        )
-        events = np.zeros(
-            shape=(
-                self.num_realizations,
-                num_steps,
-                rume.scope.nodes,
-                rume.ipm.num_events,
-            ),
-            dtype=np.int64,
-        )
-        estimated_params = {
-            name: np.zeros(
-                shape=(self.num_realizations, rume.time_frame.days, rume.scope.nodes),
-                dtype=np.float64,
-            )
-            for name in self.unknown_params.keys()
-        }
-
-        # Initialize the initial compartment and parameter values. Note, the values
-        # within current_compartment_values will be modified in place!
-        current_compartments, current_params = _initialize_compartments_and_params(
-            rume=rume,
-            unknown_params=self.unknown_params,
-            num_realizations=self.num_realizations,
-            initial_values=self.initial_values,
-            rng=rng,
-        )
-
-        initial = current_compartments.copy()
-
-        # Determine the number of observations and their associated time frames.
-        time_frames, labels = self._observation_time_frames(
-            rume, self.observations.model_link.time.grouping
-        )
-
-        context = Context.of(
-            scope=rume.scope, time_frame=rume.time_frame, ipm=rume.ipm, rng=rng
-        )
-        observations_array = self.observations._get_observations_array(context)
-        observations_array = observations_array.reshape(
-            (observations_array.shape[0], observations_array.shape[1], -1)
-        )
-
-        posterior_values = []
-
-        filter_context = self._initialize_filter_context()
-
-        day_idx = 0
-        step_idx = 0
-        for i_observation in range(len(time_frames)):
-            time_frame = time_frames[i_observation]
-
-            print(  # noqa: T201
-                (
-                    f"Observation: {i_observation}, "
-                    f"Label: {labels[i_observation]}, "
-                    f"Time Frame: {time_frame}"
-                )
-            )
-
-            result = _simulate_realizations(
-                rume_template=rume,
-                override_time_frame=time_frame,
-                num_realizations=self.num_realizations,
-                initial_values=current_compartments,
-                unknown_params=self.unknown_params,
-                param_values=current_params,
-                geo=self.observations.model_link.geo,
-                time=self.observations.model_link.time,
-                quantity=self.observations.model_link.quantity,
-                rng=rng,
-            )
-
-            day_right = day_idx + time_frame.days
-            step_right = step_idx + time_frame.days * rume.num_tau_steps
-            compartments[:, step_idx:step_right, ...] = result.compartments
-            events[:, step_idx:step_right, ...] = result.events
-
-            current_params = {}
-            for name in estimated_params.keys():
-                estimated_params[name][:, day_idx:day_right, ...] = (
-                    result.estimated_params[name]
-                )
-                current_params[name] = result.estimated_params[name][:, -1, ...].copy()
-
-            current_compartments = result.compartments[:, -1, ...].copy()
-            current_predictions = np.stack(result.predictions).reshape(
-                (self.num_realizations, rume.scope.nodes, -1)
-            )
-
-            start = np.datetime64(time_frame.start_date.isoformat())
-            end = np.datetime64(time_frame.end_date.isoformat())
-            time_frame_mask = (observations_array["date"][:, 0, 0] >= start) & (
-                observations_array["date"][:, 0, 0] <= end
-            )
-
-            current_observation = observations_array[time_frame_mask, :].reshape(
-                (rume.scope.nodes, -1)
-            )
-
-            filter_update_result = self._update(
-                current_observation=current_observation,
-                current_predictions=current_predictions,
-                prior_compartments=current_compartments,
-                prior_params=current_params,
-                filter_context=filter_context,
-                rng=rng,
-            )
-
-            current_compartments = filter_update_result.current_compartments
-            current_params = filter_update_result.current_params
-            posterior_value = filter_update_result.posterior_values
-            posterior_values.append(posterior_value)
-            filter_context = filter_update_result.filter_context
-
-            day_idx = day_right
-            step_idx = step_right
-
-        return _RunFilterResult(
-            final_compartments=current_compartments,
-            final_params=current_params,
-            compartments=compartments,
-            events=events,
-            initial=initial,
-            posterior_values=np.stack(posterior_values, axis=1),
-            estimated_params=estimated_params,
-            filter_context=filter_context,
-        )
-
-    @staticmethod
-    def _observation_time_frames(
-        rume: RUME, time_grouping: TimeGrouping | None
-    ) -> tuple[Sequence[TimeFrame], NDArray[GroupKeyType]]:
-        """
-        Find the sub time frames which correspond to each observation.
-
-        Parameters
-        ----------
-        rume :
-            The rume of the simulation.
-        time_agg :
-            The time aggregation.
-
-
-        Returns
-        -------
-        :
-            The list of time frames and the corresponding labels.
-        """
-        dim = Dim(nodes=1, days=rume.time_frame.days, tau_steps=rume.num_tau_steps)
-        dates = np.repeat(rume.time_frame.to_numpy(), rume.num_tau_steps)
-        ticks = np.arange(rume.num_ticks)
-        if time_grouping is None:
-            labels = np.broadcast_to(np.array(["*"]), ticks.shape)
-        else:
-            labels = time_grouping.map(dim, ticks, dates)
-
-        # It is important that order is preserved.
-        _, label_idx = np.unique(labels, return_index=True)
-        label_idx = np.sort(label_idx)
-        values = labels[label_idx]
-
-        time_frames = []
-        for i in range(len(label_idx) - 1):
-            start_date = datetime.date.fromisoformat(str(dates[label_idx[i]]))
-            end_date = datetime.date.fromisoformat(str(dates[label_idx[i + 1]]))
-            time_frame = TimeFrame.rangex(
-                start_date=start_date, end_date_exclusive=end_date
-            )
-            time_frames.append(time_frame)
-
-        start_date = datetime.date.fromisoformat(str(dates[label_idx[-1]]))
-        end_date = datetime.date.fromisoformat(str(dates[-1]))
-        time_frames.append(TimeFrame.range(start_date=start_date, end_date=end_date))
-
-        return time_frames, values
-
-
-###################
-# Particle Filter #
-###################
-
-
-@dataclass(frozen=True)
-class _ParticleFilterContext:
-    effective_sample_size: np.ndarray
-
-
-@dataclass(frozen=True)
-class ParticleFilterOutput(FilterOutput):
-    """
-    Output object for the particle filter which contains additional output and
-    diagnostic information.
-
-    Parameters
-    ----------
-    simulator :
-        The simulator used to produce the output.
-    posterior_values :
-        The posterior estimate of the observed data. The first dimension of the array is
-        the number of realizations, the second dimension is the number of observations.
-        The remaining dimensions depend on the shape of the observed data. These values
-        are constructed from resampling the observed data in the same way the
-        compartment and parameter values are resampled.
-    effective_sample_size :
-        The effective sample size, defined as the the reciprocal of the sum of the
-        squared weights, for each observation and each node.
-    """
-
-    simulator: "ParticleFilterSimulator"
-
-    effective_sample_size: NDArray[np.float64]
-    """
-    The effective sample size for each observation and each node.
-    """
-
-
-@dataclass(frozen=True)
-class ParticleFilterSimulator(FilterSimulator[_ParticleFilterContext]):
-    """
-    A PipelineSimulator for using a particle filter to estimate the state and parameters
-    of a system based on observed data.
-
-    Parameters
-    ----------
-    config :
-        The RUME, number of realization, initial (prior) compartment values, and unknown
-        parameters to estimate.
-    observations :
-        The observations used to estimate the compartment and parameter values.
-    """
-
-    config: PipelineConfig
-    """The configuration for the simulation."""
-
-    observations: Observations
-    """The observations."""
-
-    @staticmethod
-    def _normalize_log_weights(log_weights: NDArray[np.float64]) -> NDArray[np.float64]:
-        """
-        Calculate the normalized particle weights from the un-normalized natural
-        logarithm of the weights. This method uses multiple techniques to mitigate
-        underflow issues caused by large population sizes and multi-node systems.
-
-        Parameters
-        ----------
-        log_weights :
-            The natural logarithm of the weights.
-
-        Returns
-        -------
-        :
-            The normalized (non-log) weights.
-        """
-        # Converting unnormalized log weights to normalized non-log weights is
-        # essentially a soft max of the of the log weights.
-
-        # Shifting the log weights, equivalent to multiplying the unnormalized non-log
-        # weights by a constant, helps avoid underflow issues.
-        shifted_log_weights = log_weights - np.max(log_weights, keepdims=True)
-        underflow_lower_bound = -(10**2)
-        clipped_log_weights = np.clip(
-            shifted_log_weights, a_min=underflow_lower_bound, a_max=None
-        )
-        # Perform the softmax.
-        weights = np.exp(clipped_log_weights)
-        weights = weights / np.sum(weights, keepdims=True)
-        return weights
-
-    @staticmethod
-    def _systematic_resampling(
-        weights: NDArray, rng: np.random.Generator
-    ) -> NDArray[np.int64]:
-        """
-        Performs systematic resampling as part of a particle filter update. It is used
-        to take a weighted particle cloud and produce an equally-weighted particle cloud
-        through resampling. Each particle in the resampled cloud is a duplicate of a
-        particle in the orginal cloud. Not all particles are retained in the resampled
-        cloud. An important property of systematic resampling is that a cloud of
-        equally-weighted particles will produce the exact same cloud.
-
-
-        Parameters
-        ----------
-        weights :
-            An array of shape (R,) where R is the number of particles containing the
-            particle weights.
-        rng:
-            The random number generator.
-
-        Returns:
-        :
-            An array of shape (R,) where R is the number of particles containing the
-            indices of the prior particles corresponding to the resampled particles.
-        """
-        num_realizations = len(weights)
-        c = np.cumsum(weights)
-        u = rng.uniform(0, 1)
-        resampled_idx = []
-        for i_realization in range(num_realizations):
-            u_i = (u + i_realization) / num_realizations
-            j = np.where(c >= u_i)[0][0]
-            resampled_idx.append(j)
-        return np.array(resampled_idx)
-
-    def _initialize_filter_context(self):
-        return _ParticleFilterContext(
-            effective_sample_size=np.zeros(shape=(0, self.rume.scope.nodes))
-        )
-
-    def _update(
-        self,
-        current_observation: NDArray,
-        current_predictions: NDArray,
-        prior_compartments: NDArray,
-        prior_params: Mapping[NamePattern, NDArray],
-        filter_context: _ParticleFilterContext,
-        rng: np.random.Generator,
-    ) -> _FilterUpdateResult:
-        # posterior_value will be modified in-place.
-        posterior_value = current_predictions.copy()
-
-        current_compartments = prior_compartments.copy()
-        current_params = {k: prior_params[k].copy() for k in prior_params.keys()}
-
-        posterior_value = current_predictions.copy()
-        prior_compartment_values = current_compartments.copy()
-
-        effective_sample_size = np.zeros(shape=(1, self.rume.scope.nodes))
-
-        # Hard code localization
-        for i_node in range(self.rume.scope.nodes):
-            observation_is_missing = (current_observation.size == 0) or np.ma.is_masked(
-                current_observation["value"][i_node, ...]
-            )
-
-            if observation_is_missing:
-                effective_sample_size[0, i_node] = 1.0
-            else:
-                log_likelihoods = self.observations.likelihood.compute_log(
-                    current_observation["value"][i_node, ...],
-                    current_predictions[:, i_node, ...],
-                )
-                weights = self._normalize_log_weights(log_likelihoods)
-
-                resampled_idx = self._systematic_resampling(weights, rng)
-
-                orig_idx = np.arange(0, self.num_realizations)
-                current_compartments[orig_idx, i_node, ...] = prior_compartment_values[
-                    resampled_idx, i_node, ...
-                ]
-                for name in current_params.keys():
-                    current_params[name][orig_idx, i_node] = current_params[name][
-                        resampled_idx, i_node
-                    ]
-
-                posterior_value[orig_idx, i_node, ...] = posterior_value[
-                    resampled_idx, i_node, ...
-                ]
-                effective_sample_size[0, i_node] = 1 / np.sum(weights**2)
-        return _FilterUpdateResult(
-            current_compartments=current_compartments,
-            current_params=current_params,
-            posterior_values=posterior_value,
-            filter_context=_ParticleFilterContext(
-                effective_sample_size=np.concatenate(
-                    [filter_context.effective_sample_size, effective_sample_size],
-                    axis=0,
-                )
-            ),
-        )
-
-    @override
-    def run(self, rng):
-        run_filter_result = self._run_filter(rng=rng)
-        return ParticleFilterOutput(
-            simulator=self,
-            final_compartments=run_filter_result.final_compartments,
-            final_params=run_filter_result.final_params,
-            compartments=run_filter_result.compartments,
-            events=run_filter_result.events,
-            initial=run_filter_result.initial,
-            estimated_params=run_filter_result.estimated_params,
-            posterior_values=run_filter_result.posterior_values,
-            effective_sample_size=run_filter_result.filter_context.effective_sample_size,
-        )
-
-
-##########################
-# Ensemble Kalman Filter #
-##########################
-
-
-@dataclass(frozen=True)
-class _EnsembleKalmanFilterContext: ...
-
-
-@dataclass(frozen=True)
-class EnsembleKalmanFilterOutput(FilterOutput):
-    """
-    Output object for the ensemble Kalman filter which contains additional output and
-    diagnostic information.
-
-    Parameters
-    ----------
-    simulator :
-        The simulator used to produce the output.
-    posterior_values :
-        The posterior estimate of the observed data. The first dimension of the array is
-        the number of realizations, the second dimension is the number of observations.
-        The remaining dimensions depend on the shape of the observed data. These values
-        are constructed updating the observed data in the same way the compartment and
-        parameter values are updated.
-    """
-
-    simulator: "EnsembleKalmanFilterSimulator"
-    """The simulator used to produce the output."""
-
-    posterior_values: NDArray[np.float64]
-    """The posterior estimate of the observed data."""
-
-
-@dataclass(frozen=True)
-class EnsembleKalmanFilterSimulator(FilterSimulator[_EnsembleKalmanFilterContext]):
-    """
-    A PipelineSimulator for using an ensemble Kalman filter to estimate the state and
-    parameters of a system based on observed data.
-
-    Parameters
-    ----------
-    config : PipelineConfig
-        The RUME, number of realization, initial (prior) compartment values, and unknown
-        parameters to estimate.
-    observations : Observations
-        The observations used to estimate the compartment and parameter values.
-    """
-
-    config: PipelineConfig
-    observations: Observations
-
-    def _initialize_filter_context(self):
-        return _EnsembleKalmanFilterContext()
-
-    def _update(
-        self,
-        current_observation: NDArray,
-        current_predictions: NDArray,
-        prior_compartments: NDArray,
-        prior_params: Mapping[NamePattern, NDArray],
-        filter_context: _EnsembleKalmanFilterContext | None,
-        rng: np.random.Generator,
-    ) -> _FilterUpdateResult:
-        posterior_value = current_predictions.copy().astype(np.float64)
-
-        current_compartments = prior_compartments.copy()
-        current_params = {k: prior_params[k].copy() for k in prior_params.keys()}
-        # Hard code localization
-        for i_node in range(self.rume.scope.nodes):
-            observation_is_missing = (current_observation.size == 0) or np.ma.is_masked(
-                current_observation["value"][i_node, ...]
-            )
-
-            if not observation_is_missing:
-                prior_compartments_mean = np.mean(
-                    current_compartments[:, i_node, :], axis=0
-                )
-                prior_params_mean = {
-                    k: np.mean(current_params[k][:, i_node])
-                    for k in current_params.keys()
-                }
-
-                # The double use of the term perturbation is unfortunate.
-                compartments_perturbation = (
-                    # Relies on numpy broadcasting rules, i.e. left padding shape.
-                    current_compartments[:, i_node, :] - prior_compartments_mean
-                )
-                params_perturbation = {
-                    k: current_params[k][:, i_node] - prior_params_mean[k]
-                    for k in current_params.keys()
-                }
-
-                simulated_observations = self.observations.likelihood.sample(
-                    current_predictions[:, i_node, ...], rng=rng
-                )
-                simulated_perturbations = (
-                    simulated_observations
-                    - simulated_observations.mean(axis=0, keepdims=True)
-                )
-
-                prediction_perturbations = current_predictions[
-                    :, i_node, ...
-                ] - current_predictions[:, i_node, ...].mean(axis=0, keepdims=True)
-
-                residual_cov = (
-                    1
-                    / (self.num_realizations - 1)
-                    * np.matmul(simulated_perturbations.T, simulated_perturbations)
-                )
-
-                residual_cov_inverse = np.linalg.pinv(residual_cov)
-
-                # np.matmul behaves nicely with 1-D arrays.
-                # fmt: off
-                kalman_gain_compartments = 1 / (self.num_realizations - 1) * np.matmul(compartments_perturbation.T, simulated_perturbations) @ residual_cov_inverse  # noqa: E501
-                kalman_gain_params = {
-                    k: 1 / (self.num_realizations - 1) * np.matmul(params_perturbation[k], simulated_perturbations) @ residual_cov_inverse for k in params_perturbation.keys() # noqa: E501
-                }
-                kalman_gain_prediction = 1 / (self.num_realizations - 1) * np.matmul(prediction_perturbations.T, simulated_perturbations) @ residual_cov_inverse # noqa: E501
-                # fmt: on
-
-                observation = current_observation["value"][i_node, ...]
-                for i_realization in range(self.num_realizations):
-                    innovation = (
-                        observation - simulated_observations[i_realization, ...]
-                    ).reshape((-1, 1))
-
-                    # fmt: off
-                    current_compartments[i_realization, i_node, :] += np.rint(kalman_gain_compartments @ innovation).astype(np.int64).squeeze() # noqa: E501
-                    for name in kalman_gain_params.keys():
-                        current_params[name][i_realization, i_node] += (kalman_gain_params[name] @ innovation).squeeze()  # noqa: E501
-
-                    posterior_value[i_realization, i_node, ...] += (kalman_gain_prediction @ innovation).squeeze()  # noqa: E501
-                    # fmt: on
-                current_compartments[:, i_node, :] = np.clip(
-                    current_compartments[:, i_node, :], a_min=0, a_max=None
-                )
-
-                posterior_value[:, i_node, ...] = np.clip(
-                    posterior_value[:, i_node, ...], a_min=0, a_max=None
-                )
-        return _FilterUpdateResult(
-            current_compartments=current_compartments,
-            current_params=current_params,
-            posterior_values=posterior_value,
-            filter_context=_EnsembleKalmanFilterContext(),
-        )
-
-    @override
-    def run(self, rng):
-        run_filter_result = self._run_filter(rng=rng)
-        return EnsembleKalmanFilterOutput(
-            simulator=self,
-            final_compartments=run_filter_result.final_compartments,
-            final_params=run_filter_result.final_params,
-            compartments=run_filter_result.compartments,
-            events=run_filter_result.events,
-            initial=run_filter_result.initial,
-            estimated_params=run_filter_result.estimated_params,
-            posterior_values=run_filter_result.posterior_values,
-        )
 
 
 def munge_pipeline_output(
